Add live playback options

+++ mousetracks2/gui/utils.py
@@ -25,7 +25,7 @@ def format_distance(pixels: float, ppi: float = 96.0) -> str:
     return f'{round(cm, 3)} cm'
 
 
-def format_ticks(ticks: float, ups: int = UPDATES_PER_SECOND, accuracy: int = 1, length: int = 5) -> str:
+def format_ticks(ticks: float, ups: float = UPDATES_PER_SECOND, accuracy: int = 1, length: int = 5) -> str:
     """Convert ticks to a formatted time string."""
     seconds = ticks / ups
     minutes, seconds = divmod(seconds, 60)
